modify the way to sync data

+++ src/base/AFDefine.hpp
@@ -69,10 +69,10 @@ class TABLE_EVENT_DATA
 public:
     TABLE_EVENT_DATA() = default;
 
-    uint8_t op_type_{0u};
-    size_t row_{0u};
-    uint32_t data_index_{0u};
-    uint32_t table_index_{0u};
+    uint8_t op_type_{0};
+    uint32_t row_{0};
+    uint32_t data_index_{0};
+    uint32_t table_index_{0};
     std::string table_name_{NULL_STR};
 };
